Move random data source to an example.
